--- sdk/storage/azure_storage_queue/samples/helpers/mod.rs
@@ -1,3 +1,0 @@
-pub mod endpoint;
-pub mod logs;
-pub mod random_queue_name;
